sails-app/src/requirements/tohccbase.txt
Django>=1.6,<1.7
South==0.8
django-braces==1.2
python-dateutil==2.2
django-fluent-comments>=0.9,<1.0
django-absolute==0.3
django-model-utils==1.5
django-mptt==0.6
django-stronghold==0.2
djangorestframework==2.3
django-paintstore==0.2
django-mptt-admin==0.1
django-admin-sortable2==0.2
docutils==0.11
logutils
pyreadline==2.0
six==1.4
django-mssql>=1.6,<1.7
django.js==0.8.1
django-listable==0.1.3
django-import-export
django-coverage==1.2.4
django-form-utils
MySQL-python
